pyproject.toml
[tool.poetry]
name = "pipelines"
version = "0.1.0"
description = ""
authors = ["Arina <arrishok@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
psycopg2-binary = "^2.9.5"
sqlalchemy = "^2.0.4"
dotenv = "^0.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
